Rename gateway to proxy throughout the codebase
