+++ docker-compose.yml
@@ -1,13 +1,13 @@
 services:
-  gateway-1:
-    image: 'getoptimum/gateway:latest'
+  proxy-1:
+    image: 'getoptimum/proxy:latest'
     platform: linux/amd64  
     ports:
-      - "8081:8080" # HTTP Port for the gateway
-      - "50051:50051" # gRPC Port for the gateway
+      - "8081:8080" # HTTP Port for the proxy
+      - "50051:50051" # gRPC Port for the proxy
     environment:
-      - GATEWAY_PORT=:8080
-      - CLUSTER_ID=gateway-1
+      - PROXY_PORT=:8080
+      - CLUSTER_ID=proxy-1
       - ENABLE_AUTH=false # not using Auth0 for now if needed they pass AUTH0_DOMAIN and AUTH0_AUDIENCE
       - LOG_LEVEL=debug
       - P2P_NODES=p2pnode-1:33212,p2pnode-2:33212,p2pnode-3:33212,p2pnode-4:33212 # comma separated list of p2p node
@@ -19,15 +19,15 @@ services:
       - p2pnode-2
       - p2pnode-3
       - p2pnode-4
-  gateway-2:
-    image: 'getoptimum/gateway:latest'
+  proxy-2:
+    image: 'getoptimum/proxy:latest'
     platform: linux/amd64
     ports:
-      - "8082:8080" # HTTP Port for the gateway
-      - "50052:50051" # gRPC Port for the gateway
+      - "8082:8080" # HTTP Port for the proxy
+      - "50052:50051" # gRPC Port for the proxy
     environment:
-      - GATEWAY_PORT=:8080
-      - CLUSTER_ID=gateway-1
+      - PROXY_PORT=:8080
+      - CLUSTER_ID=proxy-1
       - ENABLE_AUTH=false # not using Auth0
       - LOG_LEVEL=debug
       - P2P_NODES=p2pnode-1:33212,p2pnode-2:33212,p2pnode-3:33212,p2pnode-4:33212
@@ -49,7 +49,7 @@ services:
       - LOG_LEVEL=debug
       - CLUSTER_ID=p2pnode-1
       - NODE_MODE=optimum # options: "gossipsub", "optimum"
-      - SIDECAR_PORT=33212 # default port for sidecar is 33212 for the grpc bidirectional streaming for the gateway
+      - SIDECAR_PORT=33212 # default port for sidecar is 33212 for the grpc bidirectional streaming for the proxy
       - API_PORT=9090 # default port for API is 9090
       - IDENTITY_DIR=/identity
       # configuration for gossipsub
