--- grpc_gateway_client/.gitignore
@@ -1,3 +1,0 @@
-.env
-
-gateway-client
