+++ grpc_proxy_client/proto/proxy_stream.proto
@@ -1,0 +1,20 @@
+syntax = "proto3";
+
+package proto;
+
+option go_package = "optimum-proxy/proto;proto";
+
+// ProxyStream establishes a stream connection to send and receive messages
+// between the Optimum Proxy and the clients.
+service ProxyStream {
+  rpc ClientStream (stream ProxyMessage) returns (stream ProxyMessage);
+}
+
+// ProxyMessage represents a message sent between the Optimum Proxy and clients.
+message ProxyMessage {
+  string client_id = 1;
+  bytes message = 2;
+  string topic = 3;
+  string message_id = 4;
+  string type = 5; 
+}
